updating git version and improving logging and moved types to root
