--- app/types/mail.ts
@@ -1,7 +1,0 @@
-export type Mail = {
-	from?: string;
-	to?: string;
-	subject: string;
-	html: string;
-	text: string;
-};
