+++ package.json
@@ -43,7 +43,7 @@
     "eslint": "8.48.0",
     "eslint-config-next": "13.4.19",
     "lucide-react": "^0.274.0",
-    "next": "13.4.19",
+    "next": "latest",
     "next-themes": "^0.2.1",
     "nodemailer": "^6.9.5",
     "postcss": "8.4.29",
